feat: a new class and documentation has been added
already existing classes were moved into the created package,
a new TimestampMessageDecorator class was also created,
and also moved into the package created for it.
In addition, documentation was written to describe
the work of TimestampMessageDecorator and ConsolePrinter.
